Omaha #3410 changed lon/lat fields to floats in sfc and air obs location

Former-commit-id: a7d17bd490f2f0931f3c22ee5484214041a34bd6 [formerly aa991c922354f0a67fac08fee456ca60a30de968] [formerly a7d17bd490f2f0931f3c22ee5484214041a34bd6 [formerly aa991c922354f0a67fac08fee456ca60a30de968] [formerly 74a6c3dd20898e9bd4484376ea4e4e471bdec9af [formerly cdb0a72fcb2345564f2b944fdecd7869fd498767]]]
Former-commit-id: 74a6c3dd20898e9bd4484376ea4e4e471bdec9af
Former-commit-id: 5977b049c1eda298ce206a8d5fc34b8f235639c5 [formerly 10ea02b9366284a7b97f57ffdb7ac54688cfa0c4]
Former-commit-id: 053cf978a7bcf69a7c3cc61bc2be8386791b90aa

+++ edexOsgi/com.raytheon.uf.common.dataplugin.ldadmesonet/src/com/raytheon/uf/common/dataplugin/ldadmesonet/LdadmesonetPointDataTransform.java
@@ -43,6 +43,7 @@ import com.raytheon.uf.common.pointdata.spatial.SurfaceObsLocation;
  * ------------ ----------  ----------- --------------------------
  * Oct 09, 2009 DR2814      vkorolev    Initial creation
  * Aug 30, 2013 2298        rjpeter     Make getPluginName abstract
+ * Jul 23, 2014 3410        bclement    location changed to floats
  * 
  * </pre>
  * 
@@ -477,8 +478,8 @@ public class LdadmesonetPointDataTransform {
         mr.setDewpoint(pdv.getNumber(DEWPOINT).floatValue());
         SurfaceObsLocation loc = new SurfaceObsLocation(
                 pdv.getString(STATION_ID));
-        Double lat = pdv.getNumber(LATITUDE).doubleValue();
-        Double lon = pdv.getNumber(LONGITUDE).doubleValue();
+        float lat = pdv.getNumber(LATITUDE).floatValue();
+        float lon = pdv.getNumber(LONGITUDE).floatValue();
         loc.assignLocation(lat, lon);
         loc.setElevation(pdv.getNumber(ELEVATION).intValue());
         mr.setLocation(loc);
